+++ packages/node-core/src/index.ts
@@ -1,6 +1,9 @@
 import * as logger from './logs/exports';
 
 export { httpIntegration } from './integrations/http';
+export { httpServerSpansIntegration } from './integrations/http/httpServerSpansIntegration';
+export { httpServerIntegration } from './integrations/http/httpServerIntegration';
+
 export {
   SentryHttpInstrumentation,
   type SentryHttpInstrumentationOptions,
