+++ packages/node-core/src/integrations/http/SentryHttpInstrumentation.ts
@@ -2,7 +2,6 @@ import type { ChannelListener } from 'node:diagnostics_channel';
 import { subscribe, unsubscribe } from 'node:diagnostics_channel';
 import type * as http from 'node:http';
 import type * as https from 'node:https';
-import type { Span } from '@opentelemetry/api';
 import { context } from '@opentelemetry/api';
 import { isTracingSuppressed } from '@opentelemetry/core';
 import type { InstrumentationConfig } from '@opentelemetry/instrumentation';
@@ -22,30 +21,11 @@ type Https = typeof https;
 
 export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
   /**
-   * Whether breadcrumbs should be recorded for requests.
+   * Whether breadcrumbs should be recorded for outgoing requests.
    *
    * @default `true`
    */
   breadcrumbs?: boolean;
-
-  /**
-   * Whether to create spans for requests or not.
-   * As of now, creates spans for incoming requests, but not outgoing requests.
-   *
-   * @default `true`
-   */
-  spans?: boolean;
-
-  /**
-   * Whether to extract the trace ID from the `sentry-trace` header for incoming requests.
-   * By default this is done by the HttpInstrumentation, but if that is not added (e.g. because tracing is disabled, ...)
-   * then this instrumentation can take over.
-   *
-   * @deprecated This is always true and the option will be removed in the future.
-   *
-   * @default `true`
-   */
-  extractIncomingTraceFromHeader?: boolean;
 
   /**
    * Whether to propagate Sentry trace headers in outgoing requests.
@@ -57,20 +37,6 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
   propagateTraceInOutgoingRequests?: boolean;
 
   /**
-   * Whether to automatically ignore common static asset requests like favicon.ico, robots.txt, etc.
-   * This helps reduce noise in your transactions.
-   *
-   * @default `true`
-   */
-  ignoreStaticAssets?: boolean;
-
-  /**
-   * If true, do not generate spans for incoming requests at all.
-   * This is used by Remix to avoid generating spans for incoming requests, as it generates its own spans.
-   */
-  disableIncomingRequestSpans?: boolean;
-
-  /**
    * Do not capture breadcrumbs for outgoing HTTP requests to URLs where the given callback returns `true`.
    * For the scope of this instrumentation, this callback only controls breadcrumb creation.
    * The same option can be passed to the top-level httpIntegration where it controls both, breadcrumb and
@@ -80,73 +46,6 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
    * @param request Contains the {@type RequestOptions} object used to make the outgoing request.
    */
   ignoreOutgoingRequests?: (url: string, request: http.RequestOptions) => boolean;
-
-  /**
-   * Do not capture spans for incoming HTTP requests to URLs where the given callback returns `true`.
-   *
-   * @param urlPath Contains the URL path and query string (if any) of the incoming request.
-   * @param request Contains the {@type IncomingMessage} object of the incoming request.
-   */
-  ignoreSpansForIncomingRequests?: (urlPath: string, request: http.IncomingMessage) => boolean;
-
-  /**
-   * Do not capture the request body for incoming HTTP requests to URLs where the given callback returns `true`.
-   * This can be useful for long running requests where the body is not needed and we want to avoid capturing it.
-   *
-   * @param url Contains the entire URL, including query string (if any), protocol, host, etc. of the incoming request.
-   * @param request Contains the {@type RequestOptions} object used to make the incoming request.
-   */
-  ignoreIncomingRequestBody?: (url: string, request: http.RequestOptions) => boolean;
-
-  /**
-   * A hook that can be used to mutate the span for incoming requests.
-   * This is triggered after the span is created, but before it is recorded.
-   */
-  incomingRequestSpanHook?: (span: Span, request: http.IncomingMessage, response: http.ServerResponse) => void;
-
-  /**
-   * Controls the maximum size of incoming HTTP request bodies attached to events.
-   *
-   * Available options:
-   * - 'none': No request bodies will be attached
-   * - 'small': Request bodies up to 1,000 bytes will be attached
-   * - 'medium': Request bodies up to 10,000 bytes will be attached (default)
-   * - 'always': Request bodies will always be attached
-   *
-   * Note that even with 'always' setting, bodies exceeding 1MB will never be attached
-   * for performance and security reasons.
-   *
-   * @default 'medium'
-   */
-  maxIncomingRequestBodySize?: 'none' | 'small' | 'medium' | 'always';
-
-  /**
-   * Whether the integration should create [Sessions](https://docs.sentry.io/product/releases/health/#sessions) for incoming requests to track the health and crash-free rate of your releases in Sentry.
-   * Read more about Release Health: https://docs.sentry.io/product/releases/health/
-   *
-   * Defaults to `true`.
-   */
-  trackIncomingRequestsAsSessions?: boolean;
-
-  /**
-   * @deprecated This is deprecated in favor of `incomingRequestSpanHook`.
-   */
-  instrumentation?: {
-    requestHook?: (span: Span, req: http.ClientRequest | http.IncomingMessage) => void;
-    responseHook?: (span: Span, response: http.IncomingMessage | http.ServerResponse) => void;
-    applyCustomAttributesOnSpan?: (
-      span: Span,
-      request: http.ClientRequest | http.IncomingMessage,
-      response: http.IncomingMessage | http.ServerResponse,
-    ) => void;
-  };
-
-  /**
-   * Number of milliseconds until sessions tracked with `trackIncomingRequestsAsSessions` will be flushed as a session aggregate.
-   *
-   * Defaults to `60000` (60s).
-   */
-  sessionFlushingDelayMS?: number;
 };
 
 /**
