+++ packages/node-core/test/integrations/httpServerSpansIntegration.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { isStaticAssetRequest } from '../../src/integrations/http/incoming-requests';
+import { isStaticAssetRequest } from '../../src/integrations/http/httpServerSpansIntegration';
 
 describe('httpIntegration', () => {
   describe('isStaticAssetRequest', () => {
